Move smart pointer classes into separated header
